+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Documentation update.
 
 ### Added
+- recall_search2: Get docids from an input file and process the docids.
 - file_input: Process docids via input file.
 - remove_processed: Removes any previous processed docids from the file_dict.
 - search_docid: Call the process_docid and check on status of docid.